requirements.txt
# 核心模型 & NLP
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.1.0

# PDF 解析
pdfplumber>=0.11.0
pillow>=10.0.0

# 科学计算
numpy>=1.23.0

# 其他依赖
tqdm>=4.65.0
python-dotenv>=1.0.0
openai>=1.0.0